getFetchData refactoring
@geekeme I've seen many new programmers don't get the `fetchData` static property thing. This commit aims to make it clear (hence the name is more verbose). Also if we move the helper in an external module we can test it better :-)
